Cleaned up a lot of implementation. - InputManager core loop should be functional (famous last words). - InputRegistry is basically all buildable from the editor (need to test if profiles are working) -- the key was adding [Tool] to ALL classes & subclasses involved in editor configurability. - Added a safety mechanism for deleting a Registry's code-generated input actions: Partial classes. One part lives safely in Input/Core/Config, the other in /generated.

+++ Input/Core/InputAction/InputAction.cs
@@ -49,7 +49,7 @@ public abstract partial class InputAction : GodotObject {
   public float PhaseStartTime { get; private set; }
   public float ElapsedSecondsInPhase => (float)(Time.GetTicksMsec() / 1000.0 - PhaseStartTime);
 
-  public virtual void Process(ReadOnlySpan<InputSample> samplesThisFrame, float delta) {
+  public virtual void Process(ReadOnlySpan<InputSample> samplesThisFrame, double delta) {
     var resultPhase = Trigger.Evaluate(samplesThisFrame, delta);
     ConsumeSamples(samplesThisFrame);
